requirements.txt
Flask
Werkzeug
pillow
opencv-python
pandas
numpy
